Replace older versions with V2 versions
