+++ config/test.exs
@@ -2,6 +2,6 @@ import Config
 
 config :redis_mutex, redis_url: "redis://localhost:6379"
 
-config :redis_mutex, RedisMutexV2Test.RedisMutexUser,
-  lock_module: RedisMutex.LockV2Mock,
+config :redis_mutex, RedisMutexTest.RedisMutexUser,
+  lock_module: RedisMutex.LockMock,
   redis_url: "redis://localhost:6379"
