+++ test/test_helper.exs
@@ -1,4 +1,4 @@
 ExUnit.configure(exclude: [:redis_dependent])
 ExUnit.start(exclude: [:skip])
 
-Mox.defmock(RedisMutex.LockV2Mock, for: RedisMutex.LockV2)
+Mox.defmock(RedisMutex.LockMock, for: RedisMutex.Lock)
